Updating to module-based dependencies for better management.
