+++ commons/src/main/java/com/fernferret/allpay/commons/IncompleteBankException.java
@@ -1,0 +1,17 @@
+package com.fernferret.allpay.commons;
+
+/**
+ * @author krinsdeath
+ */
+public class IncompleteBankException extends RuntimeException {
+    private final String message;
+
+    public IncompleteBankException(String msg) {
+        this.message = msg;
+    }
+
+    public String getLocalizedMessage() {
+        return this.message;
+    }
+
+}
