+++ pom.xml
@@ -3,9 +3,24 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>com.fernferret.allpay</groupId>
     <artifactId>AllPay</artifactId>
-    <version>12</version>
+    <version>13</version>
     <name>AllPay</name>
     <description>A payment library for Bukkit.</description>
+    <packaging>pom</packaging>
+    <modules>
+        <module>commons</module>
+        <module>fe-0.6</module>
+        <module>fe-0.7</module>
+        <module>boseconomy</module>
+        <module>econxp</module>
+        <module>realeconomy</module>
+        <module>essentialsbank</module>
+        <module>iconomy4</module>
+        <module>iconomy5</module>
+        <module>iconomy6</module>
+        <module>multicurrency</module>
+        <module>itemecon</module>
+    </modules>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
@@ -49,6 +64,7 @@
                     </archive>
                 </configuration>
             </plugin>
+            <!--
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-checkstyle-plugin</artifactId>
@@ -58,6 +74,7 @@
                     <configLocation>${project.basedir}/config/ap_checks.xml</configLocation>
                 </configuration>
             </plugin>
+            -->
         </plugins>
         <resources>
             <resource>
@@ -71,84 +88,16 @@
         <dependency>
             <groupId>org.bukkit</groupId>
             <artifactId>bukkit</artifactId>
-            <version>1.1-R3</version>
-            <type>jar</type>
-            <scope>compile</scope>
-        </dependency>
-        <!-- Bukkit Dependency -->
-
-        <!-- Start of Economy Dependencies -->
-        <dependency>
-            <groupId>cosine.boseconomy</groupId>
-            <artifactId>BOSEconomy</artifactId>
-            <version>0.6.2</version>
+            <version>1.4.7-R1.0</version>
             <type>jar</type>
             <scope>compile</scope>
         </dependency>
 
         <dependency>
-            <groupId>fr.crafter.tickleman.RealShop</groupId>
-            <artifactId>RealShop</artifactId>
-            <version>0.63</version>
-            <type>jar</type>
-            <scope>compile</scope>
+            <groupId>com.fernferret.allpay.commons</groupId>
+            <artifactId>commons</artifactId>
+            <version>1</version>
         </dependency>
-
-        <dependency>
-            <groupId>com.iCo6</groupId>
-            <artifactId>iConomy</artifactId>
-            <version>6.0</version>
-            <type>jar</type>
-            <scope>compile</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>com.iConomy</groupId>
-            <artifactId>iConomy</artifactId>
-            <version>5.0</version>
-            <type>jar</type>
-            <scope>compile</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>com.nijiko.coelho.iConomy</groupId>
-            <artifactId>iConomy</artifactId>
-            <version>4.65</version>
-            <type>jar</type>
-            <scope>compile</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>com.earth2me.essentials</groupId>
-            <artifactId>Essentials</artifactId>
-            <version>2.4.2</version>
-            <type>jar</type>
-            <scope>compile</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>me.ashtheking.currency</groupId>
-            <artifactId>MultiCurrency</artifactId>
-            <version>0.09</version>
-            <type>jar</type>
-            <scope>compile</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>ca.agnate</groupId>
-            <artifactId>EconXP</artifactId>
-            <version>1.0.0</version>
-            <type>jar</type>
-            <scope>compile</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.melonbrew</groupId>
-            <artifactId>Fe</artifactId>
-            <version>0.7</version>
-            <type>jar</type>
-        </dependency>
-        <!-- End of Economy Dependencies -->
     </dependencies>
 
     <distributionManagement>
