--- src/main/java/com/fernferret/allpay/IncompleteBankException.java
@@ -1,17 +1,0 @@
-package com.fernferret.allpay;
-
-/**
- * @author krinsdeath
- */
-public class IncompleteBankException extends RuntimeException {
-    private final String message;
-
-    public IncompleteBankException(String msg) {
-        this.message = msg;
-    }
-
-    public String getLocalizedMessage() {
-        return this.message;
-    }
-
-}
